Store: move MailChimp redux elements to mailchimp folder.
