+++ client/extensions/woocommerce/app/settings/email/mailchimp/setup-mailchimp.js
@@ -21,7 +21,7 @@ import {
 	isApiKeyCorrect,
 	isSubmittingNewsletterSetting,
 	isSubmittingStoreInfo,
-	} from 'woocommerce/state/sites/settings/email/selectors';
+	} from 'woocommerce/state/sites/settings/mailchimp/selectors';
 import KeyInputStep from './setup-steps/key-input.js';
 import LogIntoMailchimp from './setup-steps/log-into-mailchimp.js';
 import NewsletterSettings from './setup-steps/newsletter-settings.js';
@@ -34,7 +34,7 @@ import {
 	submitMailChimpStoreInfo,
 	submitMailChimpCampaignDefaults,
 	submitMailChimpNewsletterSettings
-} from 'woocommerce/state/sites/settings/email/actions.js';
+} from 'woocommerce/state/sites/settings/mailchimp/actions.js';
 
 const LOG_INTO_MAILCHIMP_STEP = 'log_into';
 const KEY_INPUT_STEP = 'key_input';
