+++ client/extensions/woocommerce/app/settings/email/mailchimp/setup-steps/newsletter-settings.js
@@ -11,13 +11,13 @@ import { map } from 'lodash';
  * Internal dependencies
  */
 import { getSelectedSiteId } from 'state/ui/selectors';
-import { isRequestingLists } from 'woocommerce/state/sites/settings/email/selectors';
+import { isRequestingLists } from 'woocommerce/state/sites/settings/mailchimp/selectors';
 import FormFieldset from 'components/forms/form-fieldset';
 import FormLabel from 'components/forms/form-label';
 import FormSelect from 'components/forms/form-select';
 import FormSettingExplanation from 'components/forms/form-setting-explanation';
 import Notice from 'components/notice';
-import QueryMailChimpLists from 'woocommerce/state/sites/settings/email/queryLists';
+import QueryMailChimpLists from 'woocommerce/state/sites/settings/mailchimp/queryLists';
 
 const NewsletterSettings = ( { storeData = {}, onChange, siteId, isRequesting, translate } ) => {
 	return (
