+++ client/extensions/woocommerce/state/sites/settings/mailchimp/querySettings.jsx
