+++ client/extensions/woocommerce/state/sites/settings/mailchimp/reducer.js
