+++ client/extensions/woocommerce/state/sites/settings/mailchimp/selectors.js
@@ -16,7 +16,7 @@ const basePath = ( siteId ) => (
 		'sites',
 		siteId,
 		'settings',
-		'email',
+		'mailchimp',
 	]
 );
