+++ client/extensions/woocommerce/state/sites/settings/mailchimp/test/actions.js
