+++ client/extensions/woocommerce/state/sites/settings/mailchimp/test/reducer.js
