+++ client/extensions/woocommerce/state/sites/settings/reducer.js
@@ -9,12 +9,12 @@ import general from './general/reducer';
 import products from './products/reducer';
 import stripeConnectAccount from './stripe-connect-account/reducer';
 import tax from './tax/reducer';
-import email from './email/reducer';
+import mailchimp from './mailchimp/reducer';
 
 export default combineReducers( {
 	general,
 	products,
 	stripeConnectAccount,
 	tax,
-	email,
+	mailchimp,
 } );
